Project/requirements.txt
google-auth-oauthlib
google-api-python-client
matplotlib
pandas
os
csv
json